feat: 引入@pear-rec/server 服务，perf: 优化打包、指定类型声明文件
