+++ packages/server/index.d.ts
@@ -1,0 +1,2 @@
+declare const initServer: () => Promise<void>;
+export default initServer;
