+++ packages/server/index.ts
@@ -1,0 +1,3 @@
+import initServer from "./src";
+
+initServer();
