+++ packages/server/package.json
@@ -1,10 +1,11 @@
 {
   "name": "@pear-rec/server",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "description": "",
   "main": "dist/index.js",
+  "typings": "index.d.ts",
   "scripts": {
-    "dev": "nodemon src/index.ts",
+    "dev": "nodemon index.ts",
     "build": "rollup -c",
     "rebuild": "cnpm i sqlite3 electron-rebuild && electron-rebuild -f -w sqlite3",
     "preview": "node dist/index.js"
