+++ packages/server/rollup.config.mjs
@@ -4,10 +4,10 @@ import commonjs from "@rollup/plugin-commonjs";
 import json from "@rollup/plugin-json";
 
 export default {
-	input: "src/server.ts",
+	input: "src/index.ts",
 	output: {
 		file: "dist/index.js",
-		format: "cjs",
+		format: "es",
 	},
 	context: "window",
 	plugins: [
